requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
umap-learn>=0.5.0
jupyter>=1.0.0
pytest>=7.0.0
tqdm>=4.60.0
pandas>=1.3.0
seaborn>=0.11.0
